+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/MotionProfile.java
@@ -1,0 +1,6 @@
+package org.firstinspires.ftc.teamcode;
+
+// TODO: Implement
+// https://www.ctrlaltftc.com/advanced/motion-profiling
+public class MotionProfile {
+}
