+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/legacy/rightAuto.java
@@ -1,18 +1,26 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.legacy;
 
-import com.acmerobotics.dashboard.config.Config;
 import com.qualcomm.hardware.rev.RevTouchSensor;
 import com.qualcomm.hardware.sparkfun.SparkFunOTOS;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.ServoImplEx;
 
+import org.firstinspires.ftc.teamcode.ArmSubSystem;
+import org.firstinspires.ftc.teamcode.AutonomousHandler;
+import org.firstinspires.ftc.teamcode.DriveBaseSubsystem;
+import org.firstinspires.ftc.teamcode.SystemState;
+import org.firstinspires.ftc.teamcode.armPose;
+import org.firstinspires.ftc.teamcode.clawState;
+import org.firstinspires.ftc.teamcode.wristState;
+
 import java.util.HashMap;
 
-@Config
-@Autonomous(name="antAuto", group="Auto")
-public class antAuto extends OpMode {
+@Disabled
+@Autonomous(name="rightAuto", group="Auto")
+public class rightAuto extends OpMode {
     private AutonomousHandler autoHandler;
 
     @Override
@@ -42,7 +50,7 @@ public class antAuto extends OpMode {
         frontLeftMotor.setDirection(DcMotor.Direction.FORWARD);
         frontRightMotor.setDirection(DcMotor.Direction.REVERSE);
         backLeftMotor.setDirection(DcMotor.Direction.FORWARD);
-        backRightMotor.setDirection(DcMotor.Direction.REVERSE);
+        backRightMotor.setDirection(DcMotor.Direction.FORWARD);
 
         // Set motor modes
         frontLeftMotor.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
@@ -61,28 +69,59 @@ public class antAuto extends OpMode {
         // State 1
         SystemState stateOne = new SystemState();
         stateOne.clawPosition = clawState.CLOSED;
-        stateOne.wristPosition = wristState.UP;
-        stateOne.armPosition = armPose.ZERO;
-        stateOne.drivePose = new SparkFunOTOS.Pose2D(-20, 0, 0);
-        stateOne.ignorePosTime = 100000;
+        stateOne.wristPosition = wristState.DOWN;
+        stateOne.armPosition = armPose.AUTO_1;
+        stateOne.drivePose = new SparkFunOTOS.Pose2D(0, 0, 0);
+        stateOne.ignorePosTime = 1;
 
-        // State too
-        SystemState stateToo = new SystemState();
-        stateToo.clawPosition = clawState.OPENED;
-        stateToo.wristPosition = wristState.UP;
-        stateToo.armPosition = armPose.ZERO;
-        stateToo.drivePose = new SparkFunOTOS.Pose2D(20, 0, 0);
-        stateToo.ignorePosTime = 100000;
+        // State 2
+        SystemState stateTwo = new SystemState();
+        stateTwo.clawPosition = clawState.CLOSED;
+        stateTwo.wristPosition = wristState.UP;
+        stateTwo.armPosition = armPose.AUTO_1;
+        stateTwo.drivePose = new SparkFunOTOS.Pose2D(-12, 18, 180);
+        stateTwo.ignorePosTime = 3.5;
+
+        // State 3
+        SystemState stateThree = new SystemState();
+        stateThree.clawPosition = clawState.CLOSED;
+        stateThree.wristPosition = wristState.UP;
+        stateThree.armPosition = armPose.AUTO_1;
+        stateThree.drivePose = new SparkFunOTOS.Pose2D(30, 9, -135);
+        stateThree.ignorePosTime = 2.5;
+
+        // State 4
+        SystemState stateFour = new SystemState();
+        stateFour.clawPosition = clawState.CLOSED;
+        stateFour.wristPosition = wristState.DOWN;
+        stateFour.armPosition = armPose.AUTO_1;
+        stateFour.drivePose = new SparkFunOTOS.Pose2D(30, 9, -135);
+        stateFour.ignorePosTime = 0.5;
+
+        // State 5
+        SystemState stateFive = new SystemState();
+        stateFive.clawPosition = clawState.CLOSED;
+        stateFive.wristPosition = wristState.DOWN;
+        stateFive.armPosition = armPose.AUTO_1;
+        stateFive.drivePose = new SparkFunOTOS.Pose2D(15.75, 25, 0); // 27
+        stateFive.ignorePosTime = 4;
+
+        // State 6
+        SystemState endState = new SystemState();
+        endState.clawPosition = clawState.OPENED;
+        endState.wristPosition = wristState.UP;
+        endState.armPosition = armPose.ZERO;
+        endState.drivePose = new SparkFunOTOS.Pose2D(0, -100, 0);
+        endState.ignorePosTime = 10;
+
         // Generate Path
         HashMap<Integer, SystemState> instructions = new HashMap<>();
-        instructions.put(0, stateOne); // Arm Up
-        instructions.put(1, stateToo); //lik ofrword or smth
-        instructions.put(2, stateOne); // Arm Up
-        instructions.put(3, stateToo); //lik ofrword or smth
-        instructions.put(4, stateOne); // Arm Up
-        instructions.put(5, stateToo); //lik ofrword or smth
-        instructions.put(6, stateOne); // Arm Up
-
+        instructions.put(0, stateOne);
+        instructions.put(1, stateTwo);
+        instructions.put(2, stateThree);
+        instructions.put(3, stateFour);
+        instructions.put(4, stateFive);
+        instructions.put(5, endState);
 
         // ACS & DBS & Handler
         ArmSubSystem armControlSubsystem = new ArmSubSystem(armPose.ZERO, cap, spindle, lswitch, LSTop, LSLower, telemetry);
